requirements.txt
streamlit
openai
geopy
folium
streamlit-folium
haversine
plotly
llama-index
llama-index-llms-openllm
python-dotenv